+++ AppboyProject/android/app/src/main/java/com/appboyproject/MainActivity.java
@@ -72,6 +72,12 @@ public class MainActivity extends ReactActivity {
     BrazeInAppMessageManager.getInstance().setCustomInAppMessageManagerListener(new BrazeInAppMessageManagerListener());
   }
 
+  @Override 
+  public void onNewIntent(Intent intent) {
+    super.onNewIntent(intent);
+    setIntent(intent);
+  }
+
   /**
    * Returns the name of the main component registered from JavaScript.
    * This is used to schedule rendering of the component.